cylc and task updates
