+++ src/swell/tasks/get_restart.py
@@ -21,23 +21,23 @@ class GetRestart(GeosTasksRunExecutableBase):
 
     # ----------------------------------------------------------------------------------------------
 
-    def cycling_restarts(self):
+    # def cycling_restarts(self):
 
-        # Get restarts (checkpoints) from the previous cycle dir
-        # ------------------------------------------------------
-        self.logger.info('GEOS restarts will be taken from the previous cycle')
+    #     # Get restarts (checkpoints) from the previous cycle dir
+    #     # ------------------------------------------------------
+    #     self.logger.info('GEOS restarts will be taken from the previous cycle')
 
-        src = os.path.join(self.previous_cycle_dir, '*_checkpoint')
+    #     src = os.path.join(self.previous_cycle_dir, '*_checkpoint')
 
-        for filepath in list(glob.glob(src)):
-            filename = os.path.basename(filepath)
-            self.fetch_to_cycle(filepath, self.at_cycle(filename))
+    #     for filepath in list(glob.glob(src)):
+    #         filename = os.path.basename(filepath)
+    #         self.fetch_to_cycle(filepath, self.at_cycle(filename))
 
-        src = os.path.join(self.previous_cycle_dir, 'tile.bin')
-        self.fetch_to_cycle(src, self.at_cycle('tile.bin'))
+    #     src = os.path.join(self.previous_cycle_dir, 'tile.bin')
+    #     self.fetch_to_cycle(src, self.at_cycle('tile.bin'))
 
-        src = os.path.join(self.previous_cycle_dir, 'RESTART', 'MOM.res.nc')
-        self.fetch_to_cycle(src, self.at_cycle(['INPUT', 'MOM.res.nc']))
+    #     src = os.path.join(self.previous_cycle_dir, 'RESTART', 'MOM.res.nc')
+    #     self.fetch_to_cycle(src, self.at_cycle(['INPUT', 'MOM.res.nc']))
 
     # ----------------------------------------------------------------------------------------------
 
@@ -61,26 +61,27 @@ class GetRestart(GeosTasksRunExecutableBase):
                            self.rst_path, 'tile.bin')
         self.fetch_to_cycle(src, self.at_cycle('tile.bin'))
 
+        # TODO: Multiple restartss
         src = os.path.join(self.swell_static_files, 'geos', 'static',
                            self.rst_path, 'MOM.res.nc')
         self.fetch_to_cycle(src, self.at_cycle(['INPUT', 'MOM.res.nc']))
 
     # ----------------------------------------------------------------------------------------------
 
-    def rename_checkpoints(self):
+    # def rename_checkpoints(self):
 
-        # Rename _checkpoint files to _rst
-        # --------------------------------
+    #     # Rename _checkpoint files to _rst
+    #     # --------------------------------
 
-        # Move to the cycle directory
-        # ---------------------------
-        os.chdir(self.cycle_dir)
+    #     # Move to the cycle directory
+    #     # ---------------------------
+    #     os.chdir(self.cycle_dir)
 
-        self.logger.info('Renaming *_checkpoint files to *_rst')
-        try:
-            os.system('rename _checkpoint _rst *_checkpoint')
-        except Exception:
-            self.logger.abort('Renaming failed, see if checkpoint files exists')
+    #     self.logger.info('Renaming *_checkpoint files to *_rst')
+    #     try:
+    #         os.system('rename _checkpoint _rst *_checkpoint')
+    #     except Exception:
+    #         self.logger.abort('Renaming failed, see if checkpoint files exists')
 
     # ----------------------------------------------------------------------------------------------
 
@@ -94,14 +95,14 @@ class GetRestart(GeosTasksRunExecutableBase):
 
         # Obtain current and previous cycle time objects, create cycle_dir
         # ----------------------------------------------------------------
-        current_cycle = scg('current_cycle')
-        cc_dto = dt.strptime(current_cycle, "%Y%m%dT%H%M%SZ")
+        # current_cycle = scg('current_cycle')
+        # cc_dto = dt.strptime(current_cycle, "%Y%m%dT%H%M%SZ")
 
-        start_cycle_point = scg('start_cycle_point')
-        sc_dto = dt.strptime(start_cycle_point, "%Y-%m-%dT%H:%M:%SZ")
+        # start_cycle_point = scg('start_cycle_point')
+        # sc_dto = dt.strptime(start_cycle_point, "%Y-%m-%dT%H:%M:%SZ")
 
-        self.cycle_dir = scg('cycle_dir')
-        self.swell_static_files = scg('swell_static_files')
+        self.cycle_dir = self.config_get('cycle_dir')
+        self.swell_static_files = self.config_get('swell_static_files')
 
         # Create cycle_dir and INPUT
         # ----------------------------
@@ -110,20 +111,20 @@ class GetRestart(GeosTasksRunExecutableBase):
 
         # rst files folder
         # ----------------
-        self.rst_path = scg('geos_restarts_directory')
+        self.rst_path = self.config_get('geos_restarts_directory')
 
-        # Previous cycle folder name
-        # ----------------------------------
-        self.forecast_duration = scg('forecast_duration')
-        self.previous_cycle_dir = self.previous_cycle(self.cycle_dir, self.forecast_duration)
+        # # Previous cycle folder name
+        # # ----------------------------------
+        # self.forecast_duration = scg('forecast_duration')
+        # self.previous_cycle_dir = self.previous_cycle(self.cycle_dir, self.forecast_duration)
 
         # Restarts should be provided
         # --------------------------------------------------
-        if sc_dto == cc_dto:
-            self.initial_restarts()
-        else:
-            self.cycling_restarts()
-            self.rename_checkpoints()
+        # if sc_dto == cc_dto:
+        self.initial_restarts()
+        # else:
+        #     self.cycling_restarts()
+        #     self.rename_checkpoints()
 
         # TODouse rst time for cap_restart
         # self.get_rst_time()
